Saving data in local storage and move const variables to another file.

+++ src/index.js
@@ -1,49 +1,28 @@
-const confirmButton = document.getElementsByClassName('confirm')[0];
-const clearButton = document.getElementsByClassName('clear')[0];
-const email = document.getElementById('email');
-const emailError = document.getElementById('email-error');
-const country = document.getElementById('country');
-const countryError = document.getElementById('country-error');
-const zip = document.getElementById('zip');
-const zipError = document.getElementById('zip-error');
-const password = document.getElementById('password');
-const passwordError = document.getElementById('password-error');
-const passwordConfirmation = document.getElementById('password-confirmation');
-const confirmationError = document.getElementById('confirmation-error');
-const spanElements = document.getElementsByTagName('span');
+import * as constVariables from './constVariables';
 
 let activeError = false;
-
 let formStorage = [];
-var addData = (() => {
 
-    let formData = {
-        email: email.value,
-        country: country.value,
-        zipCode: zip.value,
-        password: password.value,
-        passwordConfirmation: passwordConfirmation.value,
-    }
+var addData = (() => {
 
     //email field validation
     function showEmailErrors() {
-        if(email.validity.valueMissing) {
+        if(constVariables.email.validity.valueMissing) {
             //if the field is empty display the following error message.
-            emailError.textContent = 'You need to enter an e-mail address.';
-        } else if(email.validity.typeMismatch){
+            constVariables.emailError.textContent = 'You need to enter an e-mail address.';
+        } else if(constVariables.email.validity.typeMismatch){
             //if the field doesn't contain an email address display the following error message
-            emailError.textContent = 'Entered value needs to be an e-mail address.'
+            constVariables.emailError.textContent = 'Entered value needs to be an e-mail address.'
         }
-        emailError.className = 'error active';
+        constVariables.emailError.className = 'error active';
     }
 
-    email.addEventListener('input', ()=> {
+    constVariables.email.addEventListener('input', ()=> {
         //each time the user types something, we check if the form fields are valid
-        if(email.validity.valid) {
+        if(constVariables.email.validity.valid) {
             //if the field is valid, we remove error message
-            formData.email = email.value;
-            emailError.textContent = '';
-            emailError.className = 'error';
+            constVariables.emailError.textContent = '';
+            constVariables.emailError.className = 'error';
             activeError = false;
         } else {
             //if there is still an error, show the correct error
@@ -54,23 +33,22 @@ var addData = (() => {
 
     //country field validation
     function showCountryErrors() {
-        if(country.validity.valueMissing) {
+        if(constVariables.country.validity.valueMissing) {
             //if the field is empty display the following error message
-            countryError.textContent = 'You need to enter a Country.'
-        } else if(country.validity.patternMismatch) {
+            constVariables.countryError.textContent = 'You need to enter a Country.'
+        } else if(constVariables.country.validity.patternMismatch) {
              //if the field doesn't contain only letters display the following error message
-            countryError.textContent = 'Entered value needs to be only letters.';
+             constVariables.countryError.textContent = 'Entered value needs to be only letters.';
         }
-        countryError.className = 'error active';
+        constVariables.countryError.className = 'error active';
     }
 
-    country.addEventListener('input', ()=> {
+    constVariables.country.addEventListener('input', ()=> {
         //each time the user types something, we check if the form fields are valid
-        if(country.validity.valid) {
+        if(constVariables.country.validity.valid) {
             //if the field is valid, we remove error message
-            formData.country = country.value;
-            countryError.textContent = '';
-            countryError.className = 'error';
+            constVariables.countryError.textContent = '';
+            constVariables.countryError.className = 'error';
             activeError = false;
         } else {
             //if there is still an error, show the correct error
@@ -81,23 +59,22 @@ var addData = (() => {
 
     //zip code field validation
     function showZipErrors() {
-        if(zip.validity.valueMissing) {
+        if(constVariables.zip.validity.valueMissing) {
             //if the field is empty display the following error message
-            zipError.textContent = 'You need to enter a Zip Code.'
-        } else if(zip.validity.patternMismatch) {
+            constVariables.zipError.textContent = 'You need to enter a Zip Code.'
+        } else if(constVariables.zip.validity.patternMismatch) {
             //if the field doesn't contain only letters display the following error message
-            zipError.textContent = 'Entered value needs to be only numbers.';
+            constVariables.zipError.textContent = 'Entered value needs to be only numbers.';
         }
-        zipError.className = 'error active';
+        constVariables.zipError.className = 'error active';
     }
 
-    zip.addEventListener('input', ()=> {
+    constVariables.zip.addEventListener('input', ()=> {
         //each time the user types something, we check if the form fields are valid
-        if(zip.validity.valid) {
+        if(constVariables.zip.validity.valid) {
             //if the field is valid, we remove error message
-            formData.zipCode = zip.value;
-            zipError.textContent = '';
-            zipError.className = 'error';
+            constVariables.zipError.textContent = '';
+            constVariables.zipError.className = 'error';
             activeError = false;
         } else {
             //if there is still an error, show the correct error
@@ -108,23 +85,22 @@ var addData = (() => {
 
     //password field validation
     function showPasswordErrors() {
-        if(password.validity.valueMissing) {
+        if(constVariables.password.validity.valueMissing) {
             //if the field is empty display the following error message
-            passwordError.textContent = 'You need to set a password.'
-        } else if(password.validity.patternMismatch) {
+            constVariables.passwordError.textContent = 'You need to set a password.'
+        } else if(constVariables.password.validity.patternMismatch) {
             //if the field doesn't contain only letters display the following error message
-            passwordError.textContent = 'Must contain at least one number, one uppercase and one lowercase letters, one symbol. ';
+            constVariables.passwordError.textContent = 'Must contain at least one number, one uppercase and one lowercase letters, one symbol. ';
         }
-        passwordError.className = 'error active';
+        constVariables.passwordError.className = 'error active';
     }
 
-    password.addEventListener('input', ()=> {
+    constVariables.password.addEventListener('input', ()=> {
         //each time the user types something, we check if the form fields are valid
-        if(password.validity.valid) {
+        if(constVariables.password.validity.valid) {
             //if the field is valid, we remove error message
-            formData.password = password.value;
-            passwordError.textContent = '';
-            passwordError.className = 'error';
+            constVariables.passwordError.textContent = '';
+            constVariables.passwordError.className = 'error';
             activeError = false;
         } else {
             //if there is still an error, show the correct error
@@ -134,48 +110,79 @@ var addData = (() => {
     });
 
     //password confirmation field validation
-    passwordConfirmation.addEventListener('input', ()=> {
+    function showPasswordConfirmationErrors() {
+        if(constVariables.passwordConfirmation.validity.valueMissing) {
+            //if the field is empty display the following error message
+            constVariables.confirmationError.textContent = 'You need to enter the same password again.';
+        } else if(constVariables.password.value !== constVariables.passwordConfirmation.value) {
+            //if the field doesn't contain only letters display the following error message
+            constVariables.confirmationError.textContent = 'The passwords must be the same.';
+        }
+        constVariables.confirmationError.className = 'error active';
+    }
+
+    constVariables.passwordConfirmation.addEventListener('input', ()=> {
         //each time the user types something, we check if the form fields are valid
-        if(password.value == passwordConfirmation.value) {
+        if(constVariables.password.value == constVariables.passwordConfirmation.value) {
             //if the field is valid, we remove error message
-            formData.passwordConfirmation = passwordConfirmation.value;
-            confirmationError.textContent = '';
-            confirmationError.className = 'error';
+            constVariables.confirmationError.textContent = '';
+            constVariables.confirmationError.className = 'error';
             activeError = false;
         } else {
             //if there is still an error, show the correct error
-            confirmationError.textContent = 'The passwords must be the same.';   
-            confirmationError.className = 'error active';
+            showPasswordConfirmationErrors();
             activeError = true;
         }
     });
 
-    clearButton.addEventListener('click', () => {
+    function clearForm() {
         document.forms[0].reset();
         activeError = false;
-        for(let  i = 0; i<spanElements.length; i++){
-            if(spanElements[i].className == 'error active' && spanElements[i].textContent != '') {
-                spanElements[i].className = 'error';
-                spanElements[i].textContent = '';
+        for(let  i = 0; i < constVariables.spanElements.length; i++){
+            if(constVariables.spanElements[i].className == 'error active' && constVariables.spanElements[i].textContent != '') {
+                constVariables.spanElements[i].className = 'error';
+                constVariables.spanElements[i].textContent = '';
             }
         }
+    }
+
+    constVariables.clearButton.addEventListener('click', () => {
+        clearForm();
     });
 
-    confirmButton.addEventListener('click', () => {
-        if(activeError == false) {
-            formStorage.push(formData);
-
-            document.forms[0].reset();
-            for(let  i = 0; i<spanElements.length; i++){
-                if(spanElements[i].className == 'error active' && spanElements[i].textContent != '') {
-                    spanElements[i].className = 'error';
-                    spanElements[i].textContent = '';
-                }
-            }
-
-            console.warn('added', {formStorage});
-            localStorage.setItem('MyFormData', JSON.stringify(formStorage));
+    
+    constVariables.confirmButton.addEventListener('click', () => {
+        let formData = {
+            email: constVariables.email.value,
+            country: constVariables.country.value,
+            zipCode: constVariables.zip.value,
+            password: constVariables.password.value,
+            passwordConfirmation: constVariables.passwordConfirmation.value,
+        }
+        //if there is no empty fields or active errors push data in form storage and reset the form inputs
+        if(formData.email != '' &&
+            formData.country != '' &&
+                formData.password != '' &&
+                    formData.passwordConfirmation != '' &&
+                        activeError == false) {
             
+            formData.email = constVariables.email.value;
+            formData.country = constVariables.country.value;
+            formData.zipCode = constVariables.zip.value;
+            formData.password = constVariables.password.value;
+            formData.passwordConfirmation = constVariables.passwordConfirmation.value;
+            
+            formStorage.push(formData);
+            
+            clearForm();
+            console.warn('added', {formStorage});
+            localStorage.setItem('MyFormData', JSON.stringify(formStorage)); 
+        } else {
+            showEmailErrors();
+            showCountryErrors();
+            showZipErrors();
+            showPasswordErrors();
+            showPasswordConfirmationErrors();
         }
     });
